src/main/java/org/nix/home/service/UserService.java
package org.nix.home.service;

import org.nix.home.dao.UserRepository;
import org.nix.home.pojo.entity.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import javax.annotation.Resource;

@Service
public class UserService implements UserDetailsService {

    @Resource
    private UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        return findUserByName(username);
    }

    public User findUserByName(String username) {
        return userRepository.findByUsername(username);
    }


}
